create types folder and move all types to folder to clean up code
