+++ web/frontend/src/components/FeedbackBar/FeedbackBar.tsx
@@ -2,6 +2,7 @@ import { assets } from '@/assets/icons/assets';
 import './FeedbackBar.scss';
 import { useContext, useState } from 'react';
 import { Context } from '@/context/Context';
+import { VoteType } from '@/types';
 
 // FeedbackBar component
 const FeedbackBar = () => {
@@ -13,11 +14,11 @@ const FeedbackBar = () => {
   const [activeButton, setActiveButton] = useState<string | null>(null);
 
   // Function to handle user vote
-  const handleVote = (type: 'up_vote' | 'down_vote') => {
+  const handleVote = (type: VoteType) => {
     if (activeButton === type) {
       // If the same button is clicked again, reset the vote
       setActiveButton(null);
-      sendUserFeedback?.('no_vote');
+      sendUserFeedback?.(VoteType.novote);
     } else {
       // Set the new active button and send the feedback
       setActiveButton(type);
@@ -31,17 +32,17 @@ const FeedbackBar = () => {
       <div className="feedback-buttons">
         {/* Thumbs up button */}
         <button
-          className={`thumb-button ${activeButton === 'up_vote' ? 'active' : ''}`}
+          className={`thumb-button ${activeButton === VoteType.upvote ? 'active' : ''}`}
           title="Good Response"
-          onClick={() => handleVote('up_vote')}
+          onClick={() => handleVote(VoteType.upvote)}
         >
           <img src={assets.thumbs_up} alt="Good Response" />
         </button>
         {/* Thumbs down button */}
         <button
-          className={`thumb-button ${activeButton === 'down_vote' ? 'active' : ''}`}
+          className={`thumb-button ${activeButton === VoteType.downvote ? 'active' : ''}`}
           title="Bad Response"
-          onClick={() => handleVote('down_vote')}
+          onClick={() => handleVote(VoteType.downvote)}
         >
           <img src={assets.thumbs_down} alt="Bad Response" />
         </button>
