+++ web/frontend/src/types/index.ts
@@ -1,0 +1,7 @@
+export * from './types/chat.types';
+export * from './types/context.types';
+export * from './types/modal.types';
+export * from './types/feedback.types';
+export * from './types/answerSection.types';
+export * from './types/scrollButton.types';
+export * from './types/analytics.types';
