+++ web/frontend/src/types/types/answerSection.types.ts
@@ -1,0 +1,11 @@
+import { TopKItem } from './chat.types';
+
+export interface AnswerSectionProps {
+  message: {
+    content: string;
+    topk?: TopKItem[];
+  };
+  isLastMessage: boolean;
+  generationComplete: boolean;
+  recording_id: string;
+}
