+++ web/frontend/src/types/types/modal.types.ts
@@ -1,0 +1,12 @@
+export interface ModalOption {
+  text: string;
+  onClick: () => void;
+}
+
+export interface ModalDialogProps {
+  title: string;
+  description: React.ReactNode;
+  option1?: ModalOption;
+  option2?: ModalOption;
+  closeOnOutsideClick?: boolean;
+}
