+++ web/frontend/src/types/types/scrollButton.types.ts
@@ -1,0 +1,4 @@
+export interface ScrollButtonProps {
+  scrollableElementId: string;
+  generationComplete: boolean;
+}
